WIP reconcile demo and starter
