+++ sites/__vital__/package.json
@@ -24,9 +24,7 @@
     "@bodiless/ui": "^1.0.0-beta.9",
     "@bodiless/vital-editors": "^1.0.0-beta.9",
     "@bodiless/vital-elements": "^1.0.0-beta.9",
-    "@bodiless/vital-flowcontainer": "^1.0.0-beta.9",
     "@bodiless/vital-link": "^1.0.0-beta.9",
-    "@bodiless/vital-list": "^1.0.0-beta.9",
     "@bodiless/vital-templates": "^1.0.0-beta.9",
     "@bodiless/webpack": "^1.0.0-beta.9",
     "autoprefixer": "^10.4.2",
@@ -34,15 +32,11 @@
     "dotenv": "^8.2.0",
     "esm": "^3.2.25",
     "express": "^4.17.1",
-    "find-up": "^4.1.0",
     "gatsby": "^4.11.2",
-    "gatsby-link": "^4.11.1",
     "gatsby-plugin-canonical-urls": "^4.10.0",
     "gatsby-plugin-sitemap": "^5.10.2",
-    "gatsby-react-router-scroll": "^5.11.0",
     "glob": "^7.1.6",
     "informed": "^3.34.0",
-    "isomorphic-git": "^1.10.1",
     "lodash": "^4.17.19",
     "mobx-react": "^7.2.1",
     "ora": "^4.0.2",
