+++ sites/vital-demo/gatsby-config.js
@@ -51,11 +51,10 @@ const plugins = [
         // ".bl-edit" files, enabling static replacement on all packages. False disables static
         // replacement, which is the same as passing `enabled: false` in the options object.
         //
-        // You can also pass a RegExp which will be matched against the **absolute**
-        // path of every .bl-edit file being imported. If it matches and a sibling .static
-        // file is found, the file will be replaced. For instance, if you want to include
-        // all `vital-editors` and `vital-link` files, you may pass `/vital-editors|vital-link/`
-        // to this option.
+        // You can also pass a RegExp which will be matched against the **absolute** path of every
+        // .bl-edit file being imported. If it matches and a sibling .static file is found, the file
+        // will be replaced. For instance, if you want to include all `vital-editors` and
+        // `vital-link` files, you may pass `/vital-editors|vital-link/` to this option.
         //
         // Defaults to false.
         include: true,
@@ -88,16 +87,6 @@ const plugins = [
     resolve: '--vital--',
   }
 ];
-
-if (process.env.BODILESS_BUILD_STATS === '1') {
-  plugins.push({
-    resolve: 'gatsby-plugin-webpack-bundle-analyser-v2',
-    options: {
-      analyzerMode: 'static',
-      generateStatsFile: true,
-    }
-  });
-}
 
 const robotsTxtPolicy = [
   {
