+++ sites/vital-demo/gatsby-node.js
@@ -8,7 +8,7 @@ const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
 const path = require('path');
 const fs = require('fs');
 const glob = require('glob');
-const { addTokenShadowPlugin } = require('@bodiless/webpack');
+const { addTokenShadowPlugin, addStatoscopePlugin } = require('@bodiless/webpack');
 const shadow = require('--vital--/shadow');
 
 // Fix sourcemap issue
@@ -47,5 +47,18 @@ exports.onCreateWebpackConfig = ({ stage, actions }) => {
         ),
       }
     });
+  }
+  // Always on bottom to keep the plugin as last,
+  if (stage === 'build-javascript') {
+    const options = {
+      enabled: process.env.BODILESS_BUILD_STATS === '1',
+      sitePath: path.resolve('./'),
+      name: '__vital__',
+      open: process.env.BODILESS_OPEN_STATS === '1' ? 'file' : false,
+    };
+
+    actions.setWebpackConfig(
+      addStatoscopePlugin({}, options)
+    );
   }
 };
