+++ sites/vital-demo/package.json
@@ -77,7 +77,6 @@
     "copyfiles": "^2.1.1",
     "cross-env": "^5.2.0",
     "gatsby-plugin-compile-es6-packages": "^1.1.0",
-    "gatsby-plugin-webpack-bundle-analyser-v2": "^1.1.26",
     "npm-run-all": "^4.1.5",
     "tailwindcss": "^3.0.23",
     "tailwindcss-dir": "^4.0.0",
